update proposal kernel and other kernels in FPGA v2 track fixed#1676 (#1677)
* update concat and split kernel and related files in FPGA v2(v3) track

* update

* update

* update kernel and related files in FPGA v2 track

* update

* update

* update kernel and related files for static quantization in FPGA v2 track

* update

* update feed and fetch kernel in FPGA v2 track

* update io file

* update feed fetch and softmax kernel in FPGA v2 track

* update proposal kernel and other kernels in FPGA v2 track


Former-commit-id: a10c01f70c17bd4965f5beddd6af49a821545850

+++ src/operators/kernel/fpga/V2/slice_kernel.cpp
@@ -25,7 +25,7 @@ bool SliceKernel<FPGA, float>::Init(SliceParam<FPGA>* param) {
   fpga::format_ofm(output);
   DLOG << "input: " << param->input_;
   DLOG << "output: " << param->output_;
-  if (param->input_->type() != type_id<half>()) {
+  if (param->input_->type() != type_id<int8_t>()) {
     DLOG << "wrong type";
   }
   return true;
